+++ extensions/spring/boot/stormpath-spring-security-spring-boot-starter/src/main/resources/META-INF/spring.factories
@@ -1,2 +1,1 @@
 org.springframework.boot.autoconfigure.EnableAutoConfiguration = com.stormpath.spring.boot.autoconfigure.StormpathSpringSecurityAutoConfiguration
-org.springframework.security.config.annotation.web.configurers.AbstractHttpConfigurer = com.stormpath.spring.config.StormpathWebSecurityConfigurer
